requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread
streamlit
pandas
streamlit-option-menu
plotly
